Moved animation update code to new thing file. Moved thingdef to thing.h
