+++ src/map/map.h
@@ -11,7 +11,7 @@
 
 #include <SDL2/SDL.h>
 
-#include "../animation/animation.h"
+#include "./thing.h"
 #include "./map_loading/parse/parser.h"
 
 #define MAP_W	20
@@ -36,30 +36,6 @@ struct floorcielingdef {
 
 	// If not 0, then we treat this as having an invisible wall.
 	int invisible_wall;
-};
-
-// Defines a "thing" in the world. This is for
-// static objects in the world, but can easily
-// be extended for dynamic objects and/or NPCs.
-struct thingdef {
-	// The texture to render.
-	SDL_Surface* surf;
-	struct animdef anims[100];
-	// The global position.
-	int position[2];
-	// The global rotation in degrees.
-	int rotation;
-	// The distance from the player.
-	int dist;
-
-	// 0 is no animations. 1 is frame per orientation. 2 is
-	// animations for each orientation. This is a value that
-	// is set/interpreted on a per game basis.
-	int anim_class;
-	unsigned int num_anims;
-	unsigned int curr_anim;
-
-	int type;
 };
 
 struct mapdef {
